requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.4
